requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic[email]==2.9.2
aiosmtplib==3.0.2
jinja2==3.1.4
python-multipart==0.0.12
